MUSARK-219: Refactoring service_core to use DI and adding *.sc (intellij scala worksheets) files to .gitignore.
