+++ service_core/src/main/resources/routes
@@ -3,4 +3,6 @@
 # ~~~~
 
 # Setup endpoints
-GET     /v1/my-groups               no.uio.musit.microservice.core.resource.CoreResource.getSecurityGroupsForCurrentUser
+GET     /v1/my-groups        no.uio.musit.microservice.core.resource.CoreResource.getSecurityGroupsForCurrentUser
+
+GET     /v1/foo              no.uio.musit.microservice.core.resource.CoreResource.foo
